+++ src/api/mod.rs
@@ -1,0 +1,15 @@
+// This file is part of rust-web/twig
+//
+// For the copyright and license information, please view the LICENSE
+// file that was distributed with this source code.
+
+//! The extensions API for the Twig Template Engine.
+
+pub mod lexer;
+pub mod node;
+pub mod parser;
+pub mod token;
+pub use self::lexer::Lexer;
+pub use self::parser::Parser;
+pub use self::token::Token;
+pub use self::node::Node;
